requirements.txt
datasets~=2.17.1
tqdm~=4.66.2
requests~=2.31.0
jinja2~=3.1.3
pandas~=2.2.1
python-dotenv~=1.0.1